maintain: organize the connector code
Move everything related to authenticating requests into authn.go
Move the rest of the proxy code into proxy.go
Move updateRoles to the bottom of connector.go so that the Run function
is at the top.

No logic changes, just moves
